chore: RegisterDepart 모달 컴포넌트화
