+++ src/components/types/SignUpType/RegisterModal.ts
@@ -1,0 +1,10 @@
+import {Dispatch, SetStateAction} from 'react';
+
+export type RegisterModalProps = {
+  selectedCollege: string;
+  setSelectedCollege: Dispatch<SetStateAction<string>>;
+  selectedDepartment: string;
+  setSelectedDepartment: Dispatch<SetStateAction<string>>;
+  departModalVisible: boolean;
+  setDepartModalVisible: Dispatch<SetStateAction<boolean>>;
+};
