+++ src/constants/SignUp/RegisterDepart.ts
@@ -1,6 +1,7 @@
 export const DepartModal = {
   MESSAGE_SELECT1: '단과대학과 학과를 모두 선택해주세요.',
   MESSAGE_SELECT2: '대학 소속학과를 등록해주세요.',
+  IMGALT: 'close',
 };
 
 export const DROPDOWN_COLLEGE = {
